:sparkles:: Ajout dialog pour ajout/suppression formations et sessions

+++ src/app/features/formations-list/formations-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Formation } from 'src/app/models/formation';
 import { FormationHttpService } from 'src/app/services/formation-http.service';
+import { DialogSuppressionElementComponent } from '../dialog-suppression-element/dialog-suppression-element.component';
 
 @Component({
   selector: 'app-formations-list',
@@ -8,9 +10,11 @@ import { FormationHttpService } from 'src/app/services/formation-http.service';
   styleUrls: ['./formations-list.component.css']
 })
 export class FormationsListComponent implements OnInit {
+
   formations: Formation[];
 
-  constructor(private formationService: FormationHttpService) { }
+  constructor(private formationService: FormationHttpService,
+    public dialog: MatDialog) { }
 
   tableColumns: string[] = ['titre', 'description', 'supprimer'];
 
@@ -21,10 +25,14 @@ export class FormationsListComponent implements OnInit {
   }
 
   deleteFormation(id: number) {
-    console.log("suppression");
-      this.formationService.deleteById(id).subscribe();
-      window.location.reload();
-    
+    const dialogRef = this.dialog.open(DialogSuppressionElementComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.formationService.deleteById(id).subscribe();
+        this.formations = this.formations.filter(item => item.id != id);
+      }
+    });
   }
 
 }
